tui: Restructure component modules
