+++ radicle-tui/src/ui/components/list.rs
@@ -1,0 +1,97 @@
+use tuirealm::command::{Cmd, CmdResult};
+use tuirealm::props::{AttrValue, Attribute, Props};
+use tuirealm::tui::layout::Rect;
+use tuirealm::{Frame, MockComponent, State};
+
+use crate::ui::components::label::Label;
+use crate::ui::layout;
+use crate::ui::widget::{Widget, WidgetComponent};
+
+/// A component that displays a labeled property.
+#[derive(Clone)]
+pub struct Property {
+    label: Widget<Label>,
+    divider: Widget<Label>,
+    property: Widget<Label>,
+}
+
+impl Property {
+    pub fn new(label: Widget<Label>, divider: Widget<Label>, property: Widget<Label>) -> Self {
+        Self {
+            label,
+            divider,
+            property,
+        }
+    }
+}
+
+impl WidgetComponent for Property {
+    fn view(&mut self, properties: &Props, frame: &mut Frame, area: Rect) {
+        let display = properties
+            .get_or(Attribute::Display, AttrValue::Flag(true))
+            .unwrap_flag();
+
+        if display {
+            let labels: Vec<Box<dyn MockComponent>> = vec![
+                self.label.clone().to_boxed(),
+                self.divider.clone().to_boxed(),
+                self.property.clone().to_boxed(),
+            ];
+
+            let layout = layout::h_stack(labels, area);
+            for (mut label, area) in layout {
+                label.view(frame, area);
+            }
+        }
+    }
+
+    fn state(&self) -> State {
+        State::None
+    }
+
+    fn perform(&mut self, _cmd: Cmd) -> CmdResult {
+        CmdResult::None
+    }
+}
+
+/// A component that can display lists of labeled properties
+#[derive(Default)]
+#[allow(clippy::vec_box)]
+pub struct PropertyList {
+    properties: Vec<Widget<Property>>,
+}
+
+impl PropertyList {
+    pub fn new(properties: Vec<Widget<Property>>) -> Self {
+        Self { properties }
+    }
+}
+
+impl WidgetComponent for PropertyList {
+    fn view(&mut self, properties: &Props, frame: &mut Frame, area: Rect) {
+        let display = properties
+            .get_or(Attribute::Display, AttrValue::Flag(true))
+            .unwrap_flag();
+
+        if display {
+            let properties = self
+                .properties
+                .iter()
+                .map(|property| property.clone().to_boxed() as Box<dyn MockComponent>)
+                .collect();
+
+            let layout = layout::v_stack(properties, area);
+            for (mut property, area) in layout {
+                property.view(frame, area);
+            }
+        }
+    }
+
+    fn state(&self) -> State {
+        State::None
+    }
+
+    fn perform(&mut self, _cmd: Cmd) -> CmdResult {
+        CmdResult::None
+    }
+}
